requirements.txt
absl-py==0.5.0
alabaster==0.7.10
anaconda-client==1.7.2
anaconda-navigator==1.9.12
anaconda-project==0.8.2
asgiref==3.2.5
asn1crypto==0.24.0
astor==0.7.1
astroid==1.6.3
astropy==3.0.2
attrs==19.3.0
Automat==0.7.0
autopep8==1.6.0
Babel==2.5.3
backcall==0.2.0
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.9.1
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==3.1.5
bokeh==0.12.16
boto==2.48.0
boto3 @ file:///home/conda/feedstock_root/build_artifacts/boto3_1593562743480/work
botocore @ file:///home/conda/feedstock_root/build_artifacts/botocore_1596580359836/work
Bottleneck==1.2.1
brotlipy==0.7.0
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1594141588948/work
clyent==1.2.2
colorama==0.3.9
conda==4.8.3
conda-build==3.10.5
conda-package-handling==1.7.0
conda-verify==3.4.2
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.9.2
cssselect==1.0.3
cycler==0.10.0
Cython==0.28.2
cytoolz==0.10.1
dask==0.17.5
datashape==0.5.4
dbus-python==1.2.16
decorator==4.4.2
defusedxml==0.6.0
distributed==1.21.8
Django==3.0
djangorestframework==3.12.4
docutils==0.14
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.8
future==0.18.2
gast==0.2.0
gevent==1.3.0
glob2==0.7
gmpy2==2.0.8
greenlet==0.4.13
grpcio==1.15.0
h5py==2.7.1
HeapDict==1.0.1
html5lib==1.0.1
hyperlink==19.0.0
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imageio==2.3.0
imagesize==1.0.0
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1593446433964/work
imutils==0.5.2
incremental==17.5.0
ipykernel @ file:///tmp/build/80754af9/ipykernel_1595346559376/work/dist/ipykernel-5.3.3-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1593447367857/work
ipython_genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.4
itsdangerous==1.1.0
jdcal==1.4
jedi @ file:///tmp/build/80754af9/jedi_1592841914522/work
jeepney==0.4
Jinja2==2.11.2
jmespath @ file:///home/conda/feedstock_root/build_artifacts/jmespath_1589369830981/work
joblib @ file:///tmp/build/80754af9/joblib_1594236160679/work
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1594826976318/work
jupyter-console==5.2.0
jupyter-core==4.6.3
jupyterlab==2.1.5
jupyterlab-launcher==0.10.5
jupyterlab-server @ file:///tmp/build/80754af9/jupyterlab_server_1594164409481/work
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keyring==18.0.0
kiwisolver==1.2.0
lazy-object-proxy==1.3.1
llvmlite==0.23.2+1.g67a89c7
locket==0.2.0
lxml==4.2.1
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.0
mkl-random==1.0.1
more-itertools==8.4.0
mpmath==1.1.0
msgpack==1.0.0
multipledispatch==0.6.0
mysql-connector==2.2.9
mysql-connector-python==8.0.15
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.7
networkx @ file:///tmp/build/80754af9/networkx_1594377231366/work
nltk==3.3
nose==1.3.7
notebook==6.0.3
numba==0.38.0
numexpr==2.6.5
numpy==1.14.3
numpydoc==0.8.0
odo==0.5.1
olefile==0.46
opencv-python==3.4.3.18
openpyxl==2.5.3
packaging==20.4
pandas==0.23.0
pandocfilters==1.4.2
parsel==1.5.1
parso==0.7.0
partd==1.1.0
path.py==11.0.1
pathlib2==2.3.2
patsy==0.5.0
pbr==5.4.5
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///tmp/build/80754af9/pillow_1594307298074/work
pkginfo==1.5.0.1
plotly==3.9.0
pluggy==0.13.1
ply==3.11
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.6.1
psutil==5.7.0
ptyprocess==0.6.0
py @ file:///tmp/build/80754af9/py_1593446248552/work
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.8.0
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pycrypto==2.6.1
pycurl==7.43.0.2
pydbus==0.6.0
PyDispatcher==2.0.5
pyflakes==1.6.0
Pygments==2.6.1
PyHamcrest==1.9.0
pylint==1.8.4
pyo==1.0.2
pyodbc==4.0.23
PyOpenGL==3.1.1a1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
pytest==5.4.3
pytest-arraydiff==0.2
pytest-astropy==0.3.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.5.0
pytest-remotedata==0.3.2
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==0.5.2
PyYAML==5.3.1
pyzmq==19.0.1
QtAwesome==0.5.7
qtconsole==4.3.1
QtPy==1.9.0
queuelib==1.5.0
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
retrying==1.3.3
rope==0.10.7
ruamel_yaml==0.15.87
s3transfer==0.3.3
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.1.0
Scrapy==1.6.0
seaborn==0.8.1
SecretStorage==3.1.1
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.15.0
snowballstemmer==1.2.1
sortedcollections==0.6.1
sortedcontainers==2.2.2
soupsieve==2.0.1
Sphinx==1.7.4
sphinxcontrib-websupport==1.0.1
spyder-kernels==0.4.4
SQLAlchemy @ file:///tmp/build/80754af9/sqlalchemy_1593446287777/work
sqlparse==0.3.1
statsmodels==0.9.0
stevedore==1.32.0
sympy==1.1.1
tables==3.4.3
tb-nightly==1.12.0a20180928
TBB==0.1
tblib==1.6.0
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
texttable==1.6.3
tf-nightly==1.12.0.dev20180929
Theano==1.0.3+1.gc5c549bc8
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
toml==0.10.2
toolz==0.10.0
tornado==6.0.4
tqdm @ file:///tmp/build/80754af9/tqdm_1593446365756/work
traitlets==4.3.3
Twisted==19.2.0
typing==3.6.4
unicodecsv==0.14.1
urllib3==1.25.9
virtualenv==16.2.0
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
w3lib==1.20.0
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.10.11
wurlitzer==1.0.2
wxPython==4.1.0
xlrd==1.1.0
XlsxWriter==1.0.4
xlwt==1.3.0
xmltodict==0.12.0
zict==2.0.0
zipp==3.1.0
zope.interface==4.6.0
